Adding missing sprite and trying to correct start time bug with remote server

+++ main/comp_net/include/asteroid/packet_type.h
@@ -178,7 +178,7 @@ inline sf::Packet& operator>>(sf::Packet& packet, PlayerInputPacket& playerInput
 
 struct StartGamePacket : TypedPacket<PacketType::START_GAME>
 {
-    std::array<std::uint8_t, sizeof(unsigned)> startTime{};
+    std::array<std::uint8_t, sizeof(unsigned long long)> startTime{};
 };
