+++ main/comp_net/src/asteroid_net/network_client.cpp
@@ -326,7 +326,7 @@ void ClientNetworkManager::ReceivePacket(sf::Packet& packet, PacketSource source
 	case asteroid::PacketType::START_GAME:
 	{
 		const auto* startGamePacket = static_cast<const asteroid::StartGamePacket*>(receivePacket.get());
-		gameManager_.StartGame(ConvertFromBinary<unsigned>(startGamePacket->startTime));
+		gameManager_.StartGame(ConvertFromBinary<unsigned long long>(startGamePacket->startTime));
 		break;
 	}
 	case asteroid::PacketType::JOIN_ACK:
